==> notebooks/step3_final-data.11222024.ipynb <==
import pandas as pd
import numpy as np
import pycaret
from pycaret.classification import load_model

# Load the data
# challenge_abtier_processed_nov11.csv  challenge_cell_freq_processed_nov11.csv  challenge_t_cell_processed_task4.1_nov11.csv
task1 = pd.read_csv('/home/vmottaqi/cmipb_challenge/vincent_files/challenge_abtier_processed_nov17.csv', index_col=0)
task2 = pd.read_csv('/home/vmottaqi/cmipb_challenge/vincent_files/challenge_cell_freq_processed_nov17.csv', index_col=0)
task4 = pd.read_csv('/home/vmottaqi/cmipb_challenge/vincent_files/challenge_t_cell_processed_task4.1_nov17.csv', index_col=0)

# sort df by specimen_id
task1 = task1.sort_values('subject_id')
task2 = task2.sort_values('subject_id')
task4 = task4.sort_values('subject_id')

# Load submission template
submission = pd.read_csv('/home/jrollins/cmipb-challenge/3rdChallengeSubmissionTemplate_10032024.tsv', sep='\t')

# Task 1

# Load a pycaret model /home/jrollins/cmipb-challenge/task1-2_best_scc_model.pkl
from pycaret.classification import load_model
model = load_model('/home/jrollins/cmipb-challenge/task1-2_best_scc_model')

# Use the model to make predictions
predictions = model.predict(task1)

predictions

# Multiply predictions by day_0_IgG_PT for Day 14 predictions without fold change
task11_pred = predictions * task1['day_0_IgG_PT']

task11_pred = task11_pred.to_list()

task11_pred

len(predictions)

# rank the task 1.1 predictions 1 to 54 with the highest as 1
pred_rank_11 = np.argsort(task11_pred) + 1
# rank the task 1.2 predictions 1 to 54 with the highest as 1
pred_rank_12 = np.argsort(predictions) + 1


pred_rank_11

pred_rank_12

# Load submission template from tsv
submission['1.1) IgG-PT-D14-titer-Rank'] = pred_rank_11
submission['1.2) IgG-PT-D14-FC-Rank'] = pred_rank_12

# Task 2

# load /home/jrollins/cmipb-challenge/task2-1_best_scc_model.pkl
from pycaret.classification import load_model
model = load_model('/home/jrollins/cmipb-challenge/task2-1_best_scc_model')


# Use the model to make predictions
predictions = model.predict(task2)

predictions

pred_rank_21 = np.argsort(predictions) + 1

pred_rank_21

pred22 = predictions/task2['day_0_Monocytes']

pred22

pred_rank_22 = np.argsort(pred22) + 1

pred_rank_22

# replace NaN with 'NA'
pred_rank_22 = np.where('0', 'NA', pred_rank_22)
pred_rank_22

# make all the strings ints for pred_rank_22
pred_rank_22 = pred_rank_22.astype(int)

pred_rank_22

pred_rank_21
pred_rank_22

submission['2.1) Monocytes-D1-Rank'] = pred_rank_21
submission['2.2) Monocytes-D1-FC-Rank'] = pred_rank_22

submission

# Task 4.1

# Load the Task 4.1 model
model41_th1 = load_model('/home/jrollins/cmipb-challenge/Th1_model_11102024_58scc')
model41_th2 = load_model('/home/jrollins/cmipb-challenge/Th2_model_11102024_58scc')

# get predictions for task 4.1
sel_feat = ['day_0_Th2_PHA', 'day_0_IL17A_PT', 'biological_sex', 'day_0_Th2_PT', 'day_0_IL17A_PHA', 'day_0_Th1_PHA', 'day_0_Th1_PT', 'age', 'infancy_vac']
X_41 = task4[sel_feat]

predictions_th1 = model41_th1.predict(X_41)
predictions_th2 = model41_th2.predict(X_41)

merg = predictions_th1/predictions_th2

merg

# rank the predictions 1 to 54 with the highest as 1
pred_rank_41 = np.argsort(merg) + 1

pred_rank_41

submission['4.1) IFNG/IL5-Polarization-D30-Rank'] = pred_rank_41

# Fill all NaN with "NA"
submission = submission.fillna("NA")

submission 

# save the submission to tsv
submission.to_csv('/home/jrollins/cmipb-challenge/Mottaqi_Rollins_3rdChallenge_final_11222024.tsv', sep='\t', index=False)
